+++ src/o365/commands/login.ts
@@ -34,6 +34,12 @@ class LoginCommand extends Command {
     return 'Log in to Office 365';
   }
 
+  public getTelemetryProperties(args: CommandArgs): any {
+    const telemetryProps: any = super.getTelemetryProperties(args);
+    telemetryProps.authType = args.options.authType || 'deviceCode';
+    return telemetryProps;
+  }
+
   public commandAction(cmd: CommandInstance, args: CommandArgs, cb: (err?: any) => void): void {
     const chalk: any = vorpal.chalk;